refactor: global refactor - smarter file repartition - clean code
