+++ request.go
@@ -1,37 +1,38 @@
 package main
 
 import (
-	"bytes"
-	"encoding/json"
-	"net/http"
+	"fmt"
+	abci "github.com/gnolang/gno/pkgs/bft/abci/types"
+	"github.com/gnolang/gno/pkgs/bft/rpc/client"
 )
 
-func QueueRequest(a []Embed) {
-	var b []Embed
-	for i, p := range a {
-		b = append(b, p)
-		if (i+1)%10 == 0 {
-			err := sendRequest(b)
-			if err != nil {
-				return
-			}
-			b = nil
-		}
+func makeRequest(qpath string, data []byte) (res *abci.ResponseQuery, err error) {
+	opts2 := client.ABCIQueryOptions{
+		// Height: height, XXX
+		// Prove: false, XXX
 	}
-	if len(b) > 0 {
-		err := sendRequest(b)
-		if err != nil {
-			return
-		}
+	remote := "gno.land:36657"
+	cli := client.NewHTTP(remote, "/websocket")
+	qres, err := cli.ABCIQueryWithOptions(qpath, data, opts2)
+	if err != nil {
+		return nil, err
 	}
+	if qres.Response.Error != nil {
+		fmt.Printf("Log: %s\n",
+			qres.Response.Log)
+		return nil, qres.Response.Error
+	}
+	return &qres.Response, nil
 }
 
-func sendRequest(e []Embed) error {
-	b := Body{Embeds: e}
-	jsonData, _ := json.Marshal(b)
-	_, err := http.Post(WebhookURL, "application/json", bytes.NewBuffer(jsonData))
+func getBoardsContents(board string) (string, error) {
+	qpath := "vm/qrender"
+	data := []byte(fmt.Sprintf("%s\n%s", "gno.land/r/boards", board))
+	res, err := makeRequest(qpath, data)
+
 	if err != nil {
-		return err
+		fmt.Println("Error: ", res.Log)
+		return "", err
 	}
-	return nil
+	return string(res.Data), nil
 }
